[Obs AI Assistant] Show Inspect button when no ML nodes are available (#220173)
(cherry picked from commit 1971ff261c9d2b48d29a2a5414d935d788c801cd)
