+++ x-pack/platform/packages/shared/kbn-ai-assistant/src/knowledge_base/knowledge_base_installation_status_panel.tsx
@@ -5,22 +5,14 @@
  * 2.0.
  */
 
-import React, { useState } from 'react';
-import {
-  EuiButton,
-  EuiButtonEmpty,
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiPopover,
-  EuiSpacer,
-  EuiText,
-} from '@elastic/eui';
+import React from 'react';
+import { EuiButton, EuiFlexGroup, EuiFlexItem, EuiSpacer, EuiText } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { KnowledgeBaseState } from '@kbn/observability-ai-assistant-plugin/public';
 import { UseKnowledgeBaseResult } from '../hooks';
-import { WelcomeMessageKnowledgeBaseSetupErrorPanel } from './welcome_message_knowledge_base_setup_error_panel';
 import { SelectModelAndInstallKnowledgeBase } from './select_model_and_install_knowledge_base';
 import { SettingUpKnowledgeBase } from './setting_up_knowledge_base';
+import { InspectKnowledgeBasePopover } from './inspect_knowlegde_base_popover';
 
 const WarmUpModel = ({
   knowledgeBase,
@@ -79,45 +71,6 @@ const WarmUpModel = ({
       </EuiFlexGroup>
     </>
   );
-};
-
-const InspectKnowledgeBasePopover = ({
-  knowledgeBase,
-}: {
-  knowledgeBase: UseKnowledgeBaseResult;
-}) => {
-  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
-
-  const handleInstall = async (inferenceId: string) => {
-    setIsPopoverOpen(false);
-    await knowledgeBase.install(inferenceId);
-  };
-
-  return knowledgeBase.status.value?.modelStats ? (
-    <EuiFlexItem grow={false}>
-      <EuiPopover
-        button={
-          <EuiButtonEmpty
-            data-test-subj="observabilityAiAssistantWelcomeMessageInspectErrorsButton"
-            iconType="inspect"
-            onClick={() => setIsPopoverOpen(!isPopoverOpen)}
-          >
-            {i18n.translate('xpack.aiAssistant.welcomeMessage.inspectErrorsButtonEmptyLabel', {
-              defaultMessage: 'Inspect',
-            })}
-          </EuiButtonEmpty>
-        }
-        isOpen={isPopoverOpen}
-        panelPaddingSize="none"
-        closePopover={() => setIsPopoverOpen(false)}
-      >
-        <WelcomeMessageKnowledgeBaseSetupErrorPanel
-          knowledgeBase={knowledgeBase}
-          onRetryInstall={handleInstall}
-        />
-      </EuiPopover>
-    </EuiFlexItem>
-  ) : null;
 };
 
 export const KnowledgeBaseInstallationStatusPanel = ({
